Added more  Billing.

+++ src/main/kotlin/com/nortal/dogo/terminal/Bill.kt
@@ -2,4 +2,4 @@ package com.nortal.dogo.terminal
 
 import java.util.*
 
-data class Bill(val containerId: UUID, val cost: Int)
+data class Bill(val containerId: UUID, val cost: Long)
